Simplify android implementation
